requirements.txt
argparse
pyperclip
PyCryptodome
pyinstaller